moved QueryContext to df-valve, added QueryContext support to valves
